order/requirements.txt
alembic==1.13.1
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
Django==5.0.3
django-cors-headers==4.3.1
django-mysql==4.9.0
djangorestframework==3.14.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
mysqlclient==2.2.4
pika==1.3.2
pytz==2024.1
SQLAlchemy==2.0.28
sqlparse==0.4.4
typing_extensions==4.10.0
Werkzeug==3.0.1
